+++ apps/desktop/src/renderer/components/GlobalDialogs/SketchModulesDialog/SketchModulesDialog.tsx
@@ -1,71 +1,11 @@
-import { useCallback, useEffect, useRef } from 'react'
-import { SketchModuleItem } from '@hedron-gl/engine'
-import {
-  Button,
-  Card,
-  CardActions,
-  CardBody,
-  CardContent,
-  CardDetails,
-  CardHeader,
-  CardList,
-  Dialog,
-  Panel,
-  PanelBody,
-  PanelHeader,
-  useEngineStore,
-} from '@hedron-gl/ui-core'
+import { useEffect, useRef } from 'react'
+import { CardList, Dialog, Panel, PanelBody, PanelHeader, useEngineStore } from '@hedron-gl/ui-core'
 import c from './SketchModulesDialog.module.css'
 import { useSketchesSearch } from './useSketchesSearch'
+import { SketchCard } from './SketchCard'
 import { GlobalDialogProps } from '@components/GlobalDialogs/types'
 import { useSketchModuleList } from '@components/hooks/useSketchModuleList'
 import { useSetActiveSketchId } from '@components/hooks/useSetActiveSketchId'
-
-interface SketchCardProps {
-  item: SketchModuleItem
-  closeDialog: () => void
-}
-
-export const SketchCard = ({
-  item: {
-    moduleId,
-    config: { nodes, title, description },
-  },
-  closeDialog,
-}: SketchCardProps) => {
-  const setActiveSketchId = useSetActiveSketchId()
-  const addSketch = useEngineStore((state) => state.addSketch)
-
-  const onButtonClick = useCallback(() => {
-    const id = addSketch(moduleId)
-    setActiveSketchId(id)
-    closeDialog()
-  }, [addSketch, closeDialog, moduleId, setActiveSketchId])
-
-  const numParams = nodes.filter((n) => n.nodeType === 'param').length
-  const numShots = nodes.filter((n) => n.nodeType === 'shot').length
-
-  return (
-    <Card>
-      <CardContent>
-        <CardHeader iconName="token">{title}</CardHeader>
-        <CardDetails>
-          Params: {numParams} • Shots: {numShots}
-        </CardDetails>
-        {description && (
-          <CardBody>
-            <p>{description}</p>
-          </CardBody>
-        )}
-      </CardContent>
-      <CardActions>
-        <Button onClick={onButtonClick} iconName="add">
-          Add To Scene
-        </Button>
-      </CardActions>
-    </Card>
-  )
-}
 
 export const SketchModulesDialog = ({ closeDialog }: GlobalDialogProps) => {
   const sketchModules = useSketchModuleList()
